requirements.txt
# Core
pyyaml>=6.0.1
jinja2>=3.1.2
jsonschema>=4.21.1
requests>=2.31.0

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Dev
black>=23.12.0
ruff>=0.1.9
